api.views: move landing job views to main module
